Added twisting, reflections and refractions to ray marching

+++ graphics3D/raymarching/Light.java
@@ -6,13 +6,13 @@ import graphics3D.utils.Vector;
 public class Light {
     public Vector location = new Vector(0, 0, 0);
     public Color color = new Color(0xFFFFFF);
-    public double shadowSharpness = 1;
+    public int shadowSharpness = 1;
     
     public Light() {}
     public Light(Vector location) {
         this.location = location;
     }
-    public Light(Vector location, double shadowSharpness) {
+    public Light(Vector location, int shadowSharpness) {
         this.location = location;
         this.shadowSharpness = shadowSharpness;
     }
@@ -20,7 +20,7 @@ public class Light {
         this.location = location;
         this.color = color;
     }
-    public Light(Vector location, Color color, double shadowSharpness) {
+    public Light(Vector location, Color color, int shadowSharpness) {
         this.location = location;
         this.color = color;
         this.shadowSharpness = shadowSharpness;
